Improved QrCodeOptions use of cellSize, imageSize and marginSize parameters
Also made parseColorString a public companion object method

Now you can specify cellSize only, or cellSize and marginSize, or imageSize only, or imageSize and marginSize - if marginSize is not specified, it is the same as cellSize when cellSize is given, otherwise it is 0.

You cannot make the marginSize the same as the cellSize when the imageSize is given, because the margin is needed to calculate the cellSize.