Improved QrCodeOptions use of cellSize, imageSize and marginSize parameters
Now you can specify cellSize only, or cellSize and marginSize, or imageSize only, or imageSize and marginSize - if marginSize is not specified, it is the same as cellSize.

Turns out, you can make the marginSize the same as the cellSize when the imageSize is given, because the margin is not actually needed to calculate the cellSize if the cellSize and marginSize are assumed to be the same.