Java usability of PaginationData, use of PaginationOptions, use of RunAs
PaginationOptions would be ignored if page param not passed. Now instead it is ignored only if both page and pageSize params are not passed. If only pageSize param is passed, page defaults to 1
RunAs allows setting of delegation key more than once if the value if the same. This may be needed in some cases when tests need to reload the context more than once in the same JVM process